requirements.txt
cohere
gspread
oauth2client
python-dotenv
matplotlib
